// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface ProductDocumentDataSocialCardsItem {
  /**
   * Social Card Image field in *Produits → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: product.social_cards[].social_card_image
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  social_card_image: prismic.ImageField<never>;

  /**
   * Social Card TItle field in *Produits → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: product.social_cards[].social_card_title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  social_card_title: prismic.KeyTextField;

  /**
   * Social card description field in *Produits → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: product.social_cards[].social_card_description
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  social_card_description: prismic.KeyTextField;
}

/**
 * Content for Produits documents
 */
interface ProductDocumentData {
  /**
   * Title field in *Produits*
   *
   * - **Field Type**: Title
   * - **Placeholder**: *None*
   * - **API ID Path**: product.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  title: prismic.TitleField;

  /**
   * Etsy Url field in *Produits*
   *
   * - **Field Type**: Embed
   * - **Placeholder**: *None*
   * - **API ID Path**: product.etsy_url
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#embed
   */
  etsy_url: prismic.EmbedField;

  /**
   * image field in *Produits*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: product.image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  image: prismic.ImageField<never>;

  /**
   * Price field in *Produits*
   *
   * - **Field Type**: Number
   * - **Placeholder**: *None*
   * - **API ID Path**: product.price
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#number
   */
  price: prismic.NumberField /**
   * Meta Title field in *Produits*
   *
   * - **Field Type**: Text
   * - **Placeholder**: My title for Search Engine
   * - **API ID Path**: product.meta_title
   * - **Tab**: SEO
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Produits*
   *
   * - **Field Type**: Text
   * - **Placeholder**: The description that will appear in search engine
   * - **API ID Path**: product.meta_description
   * - **Tab**: SEO
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Social Cards - Facebook & Twitter field in *Produits*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: product.social_cards[]
   * - **Tab**: SEO
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  social_cards: prismic.GroupField<
    Simplify<ProductDocumentDataSocialCardsItem>
  >;
}